I've moved the logic of rendering the image to a separate file Renderer.py
